update db (data type)

+++ db/db_model/agreement_model.go
@@ -7,7 +7,7 @@ import (
 type Agreement struct {
 	gorm.Model
 	ID           uint       `gorm:"primary_key;autoIncrement:true"`
-	InterestRate int32      `gorm:"not null"`
+	InterestRate float32    `gorm:"not null"`
 	DueIn        int32      `gorm:"not null"`
 	Addition     string     `gorm:"default:''"`
 	Contracts    []Contract `gorm:"foreignKey:AgreementId"`
